requirements.txt
Django==1.9
line-bot-sdk==1.5.0
gunicorn==19.3.0